Refactor + Bug fix for buttons resizing images
