+++ pygbase/game_state.py
@@ -1,4 +1,3 @@
-from abc import abstractmethod
 from typing import Type
 
 import pygame
